Add authority verification in finished spendings

+++ src/main/kotlin/com/finitas/domain/services/FinishedSpendingStoreService.kt
@@ -1,5 +1,6 @@
 package com.finitas.domain.services
 
+import com.finitas.config.exceptions.NotEnoughAuthorityToOperateOnUserException
 import com.finitas.domain.dto.store.*
 import com.finitas.domain.ports.*
 import java.util.*
@@ -9,10 +10,6 @@ class FinishedSpendingStoreService(
     private val reachableUsersRepository: ReachableUsersRepository,
     private val userNotifierPort: UserNotifierPort,
 ) {
-    @Deprecated("To remove")
-    suspend fun synchronizeFinishedSpendings(request: SynchronizationRequest<FinishedSpendingDto>) =
-        repository.synchronizeFinishedSpendings(request)
-
     suspend fun updateWithChangedItems(request: List<FinishedSpendingDto>, petitioner: UUID) {
         repository.updateWithChangedItems(
             IdUserWithEntities(
@@ -37,8 +34,24 @@ class FinishedSpendingStoreService(
     }
 
 
-    suspend fun getAllFinishedSpendings(idUser: UUID) = repository.getAllFinishedSpendings(idUser)
+    suspend fun getAllFinishedSpendings(requester: UUID, idUser: UUID): List<FinishedSpendingDto> {
+        val reachableUsersForDeletion = reachableUsersRepository.getUsersUnderAuthority(
+            requester,
+            authority = Authority.READ_USERS_DATA,
+        ).users
+        if (idUser !in reachableUsersForDeletion) {
+            throw NotEnoughAuthorityToOperateOnUserException(idUser)
+        }
+        return repository.getAllFinishedSpendings(idUser)
+    }
     suspend fun createFinishedSpending(requester: UUID, dto: FinishedSpendingDto) {
+        val reachableUsersForDeletion = reachableUsersRepository.getUsersUnderAuthority(
+            requester,
+            authority = Authority.MODIFY_USERS_DATA,
+        ).users
+        if (dto.idUser !in reachableUsersForDeletion) {
+            throw NotEnoughAuthorityToOperateOnUserException(dto.idUser)
+        }
         //TODO: verify requester
         //TODO: use response
         repository.createFinishedSpending(dto)
@@ -62,6 +75,13 @@ class FinishedSpendingStoreService(
 
 
     suspend fun updateFinishedSpending(requester: UUID, dto: FinishedSpendingDto) {
+        val reachableUsersForDeletion = reachableUsersRepository.getUsersUnderAuthority(
+            requester,
+            authority = Authority.MODIFY_USERS_DATA,
+        ).users
+        if (dto.idUser !in reachableUsersForDeletion) {
+            throw NotEnoughAuthorityToOperateOnUserException(dto.idUser)
+        }
         //TODO: verify requester
         //TODO: use response
         repository.updateFinishedSpending(dto)
@@ -85,6 +105,13 @@ class FinishedSpendingStoreService(
 
 
     suspend fun deleteFinishedSpending(requester: UUID, request: DeleteFinishedSpendingRequest) {
+        val reachableUsersForDeletion = reachableUsersRepository.getUsersUnderAuthority(
+            requester,
+            authority = Authority.MODIFY_USERS_DATA,
+        ).users
+        if (request.idUser !in reachableUsersForDeletion) {
+            throw NotEnoughAuthorityToOperateOnUserException(request.idUser)
+        }
         //TODO: verify requester
         //TODO: use response
         repository.deleteFinishedSpending(request)
@@ -106,5 +133,12 @@ class FinishedSpendingStoreService(
         )
     }
 
-    suspend fun fetchUsersUpdates(request: List<IdUserWithVersion>) = repository.fetchUsersUpdates(request)
+    suspend fun fetchUsersUpdates(requester: UUID, request: List<IdUserWithVersion>): List<FetchUpdatesResponse<FinishedSpendingDto>> {
+        val reachableUsers = reachableUsersRepository.getUsersUnderAuthority(
+            requester,
+            authority = Authority.READ_USERS_DATA,
+        ).users
+
+        return repository.fetchUsersUpdates(request.filter { it.idUser in reachableUsers })
+    }
 }
